+++ Models/Employee.cs
@@ -2,14 +2,19 @@ using System;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace CommunityLink.Models {
-    public class Employee {
+namespace CommunityLink.Models
+{
+    public class Employee
+    {
         [Key]
         public int EmployeeID { get; set; }
+
         [Required]
-        public int UserID { get; set; }
+        public required string UserID { get; set; }
+
         [ForeignKey("UserID")]
-        public required User User { get; set; }
+        public required CommunityUser User { get; set; }
+
         [Required]
         [StringLength(25)]
         public required string Role { get; set; }