+++ Models/LocationPhone.cs
@@ -2,17 +2,22 @@ using System;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace CommunityLink.Models {
-    public class InventoryPhone {
+namespace CommunityLink.Models
+{
+    public class LocationPhone
+    {
         [Key]
         public int PhoneID { get; set; }
+
         [StringLength(100)]
         public string? ContactName { get; set; }
+
         [Required]
         [StringLength(16)]
         public required string PhoneNumber { get; set; }
         public int LocationID { get; set; }
+
         [ForeignKey("LocationID")]
-        public required InventoryLocation InventoryLocation { get; set;}
+        public required InventoryLocation InventoryLocation { get; set; }
     }
 }